+++ Sources/SwiftKit/Extensions/String/String+HasContent.swift
@@ -1,0 +1,16 @@
+//
+//  String+HasContent.swift
+//  SwiftKit
+//
+//  Created by Daniel Saidi on 2020-06-05.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension String {
+    
+    var hasContent: Bool {
+        count > 0
+    }
+}