+++ Sources/SwiftKit/Extensions/String/String+Replace.swift
@@ -3,7 +3,7 @@
 //  SwiftKit
 //
 //  Created by Daniel Saidi on 2016-01-08.
-//  Copyright © 2016 Daniel Saidi. All rights reserved.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
 //
 //  https://danielsaidi.com/blog/2020/06/04/string-replace
 //