+++ Tests/SwiftKitTests/Extensions/Comparable+LimitTests.swift
@@ -3,7 +3,7 @@
 //  SwiftKitTests
 //
 //  Created by Daniel Saidi on 2018-10-04.
-//  Copyright © 2018 Daniel Saidi. All rights reserved.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
 //
 
 import Quick