+++ Tests/SwiftKitTests/Extensions/String/String+UrlEncodeTests.swift
@@ -3,7 +3,7 @@
 //  SwiftKitTests
 //
 //  Created by Daniel Saidi on 2016-12-12.
-//  Copyright © 2016 Daniel Saidi. All rights reserved.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
 //
 
 import Quick