+++ Release Notes.md
@@ -1,19 +1,31 @@
 # Release notes
 
+I will bump revision by revision, until SwiftKit has all functionality that it should have from iExtra. I will then bump it to `1.0.0`.
+
+
+## 0.0.4
+
+This version adds [device][Device] and [keychain][Keychain] utils.
+
+
+## 0.0.3
+
+This version adds a bunch of convenient [extensions][Extensions]. 
+
 
 ## 0.0.2
 
-This release adjusts [authentication][Authentication] service signatures.
-
-It also adds [coding][Coding] and [IoC][IoC] functionality to `SwiftKit`. 
+This version adjusts [authentication][Authentication] service signatures, as well as [coding][Coding] and [IoC][IoC] functionality  
 
 
 ## 0.0.1
 
-This release adds [authentication][Authentication] functionality to `SwiftKit`.
+This version adds [authentication][Authentication] functionality to `SwiftKit`.
 
 
 [Authentication]: Readmes/Authentication.md
 [Coding]: Readmes/Coding.md 
+[Device]: Readmes/Device.md
 [Extensions]: Readmes/Extensions.md
 [IoC]: Readmes/IoC.md
+[Keychain]: Readmes/Keychain.md